- removed some methods from protocol service - quite big refactor with moving things to job package
