--- src/Edde/Common/Upgrade/CurrentVersionException.php
@@ -1,9 +1,0 @@
-<?php
-	declare(strict_types=1);
-
-	namespace Edde\Common\Upgrade;
-
-	use Edde\Api\Upgrade\UpgradeException;
-
-	class CurrentVersionException extends UpgradeException {
-	}
