+++ src/Edde/Ext/Job/ThreadManagerConfigurator.php
@@ -1,7 +1,7 @@
 <?php
 	declare(strict_types=1);
 
-	namespace Edde\Ext\Protocol;
+	namespace Edde\Ext\Job;
 
 	use Edde\Api\Container\LazyContainerTrait;
 	use Edde\Api\Thread\IThreadManager;
@@ -15,6 +15,5 @@
 		 */
 		public function configure($instance) {
 			parent::configure($instance);
-			$instance->registerThreadHandler($this->container->create(ProtocolThreadHandler::class));
 		}
 	}
